requirements.txt
fastapi
uvicorn[standard]
motor
jinja2
python-dotenv
pydantic-settings
openai
